title_bar: Merge linux only code into `platform_linux`
